Split activity-pane.tsx into focused chart and heatmap modules
Extract chart components (DailyStackedBarChart, DailyTokenChart,
ModelBreakdown, ProjectBreakdown) into activity-charts.tsx and heatmap
components (Heatmap, ActiveHoursGrid) into activity-heatmaps.tsx.

Reduces activity-pane.tsx from 1142 lines to 370, with the extracted
files at 505 and 290 lines respectively. No behavior change.
